% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_select.R
\name{cell.select}
\alias{cell.select}
\title{Select cells}
\usage{
cell.select(data, px, py, sect.size, p.select)
}
\arguments{
\item{data}{A data frame with N rows and 4 variables (OBS_ID, x, y, MINMAX_CLUST)}

\item{px}{Optimal power for transforming x-coordinate data}

\item{py}{Optimal power for transforming y-coordinate data}

\item{sect.size}{Sector size in degrees}

\item{p.select}{Percentage of non-gated cells to be selected in each region}
}
\value{
A data frame with N rows and 9 variables (OBS_ID, x, y, MINMAX_CLUST, tx, ty, sector, sector.sel, center.sel)

tx: power transformed x

ty: power transformed y

sector: sector indicator (1 to 8)

sector.sel: selected within a sector (0=no, 1=yes)

center.sel: selected in the center region (0=no, 1=yes)
}
\description{
This function selects cells in regions of interest.
}
