% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_select.R
\name{power.select}
\alias{power.select}
\title{Tool for selecting a transformation power}
\usage{
power.select(dframe, xy_var, pstart, pstop)
}
\arguments{
\item{dframe}{Data frame with N rows and 4 variables (OBS_ID, x, y, MINMAX_CLUST)}

\item{pstart}{Minimum power}

\item{pstop}{Maximum power}

\item{xy_coord}{Use xy_coord = 2 for x-axis variable and xy_coord = 3 for  y-axis variable}
}
\value{
Data frame of powers and corresponding MC values
}
\description{
Function that computes the Medcouple (MC), a robust measure of skewness,  for a range of power transformations.
}
